Fix bug due to calo.filter
The calo.filter="edep0" in SteeringFile_IDEA_o2_v01.py did not let the
DRTubesSDAction be executed on opticalphotons which did not deposit
energy in fibers. However the SDAction must be executed on those steps
as they are counted to create the Cherenkov signal and they are killed
to speedup the simulation. Replace to calo.filter=None. Also set the
physicsList as FTFP_BERT. And the beam to 10 GeV e- shot directly into
the endcap calorimeter to monitor the time per event.