Fix overlap of DREndcapTubes and comp solenoid
The DREndcapTubes phi air staves reached the beam pipe (y=0), this was
clearly causing an overlap with the compensating solenoid. To fix this a
new variable set the starting y of the phi air staves to 22 cm from the
z-axis. This commit was checked for overlaps with a tolerance of 10 um
(fibers not included in towers) and no overlaps were detected.