Add "system" id to DREndcapTubes subdetector
Add "system" id to DREndcapTubes xml file and assign it to the highest
volume in the geometry node. While recreating the volumeID in the
DRTubesSDAction, "system" id is set as in DectDimensions_IDEA_o2_v01.xml
file. In both xml files add a comment on ids to prevent unwanted
modifications.