Comment out volID check for barrel calo
Commet out volID printout check for barrel calo in DRTubesSDAction. This
was used to check that the 64-bits volIDs recreated from the barrel
copynumbers are identical to the DD4hep ones. However marking barrel
fibers as sensitive makes the memory footprint explode (34 Gb before it
is killed on lxplus). So the test was done on a single stave.