Equal fiber radius in IDEA_o2 endcap and barrel (#545)
The fiber core radius was different in the endcap and barrel hadronic
calorimeters of IDEA_o2. This created a different response for the two.
There is no fiber difference in the latest design, so I set them equal.
Also, I am removing the case aroung barrel towers which is not yet included
in the latest calorimeter design.